requirements.txt
Flask==1.1.2
Pillow==7.2.0
numpy==1.19.1
